LayoutTests/storage/indexeddb/mozilla/open-database-null-name.html

<!DOCTYPE html>
<!--
  original test: http://mxr.mozilla.org/mozilla2.0/source/dom/indexedDB/test/test_open_empty_db.html?force=1
  license of original test:
    " Any copyright is dedicated to the Public Domain.
      http://creativecommons.org/publicdomain/zero/1.0/ "
-->
<html>
<head>
<script src="../../../fast/js/resources/js-test-pre.js"></script>
<script src="../resources/shared.js"></script>
</head>
<body>
<script>

description("Test IndexedDB: should NOT throw when opening a database with a null name");

function test()
{
    removeVendorPrefixes();

    description = "My Test Database";
    request = evalAndLog("indexedDB.open(null);");
    request.onsuccess = openSuccess;
    request.onerror = unexpectedErrorCallback;
}

function openSuccess()
{
    db = evalAndLog("db = event.target.result");
    shouldBe("db.name", "'null'");
    finishJSTest();
}


test();

</script>
<script src="../../../fast/js/resources/js-test-post.js"></script>
</body>
</html>
